refacto SortSource to own module
