+++ polars/polars-lazy/polars-pipe/src/executors/sinks/sort/mod.rs
@@ -1,5 +1,6 @@
 mod io;
 mod ooc;
 mod sink;
+mod source;
 
 pub(crate) use sink::SortSink;
